+++ pubky-homeserver/src/core/mod.rs
@@ -1,7 +1,7 @@
 mod err_if_user_is_invalid;
 mod extractors;
 mod homeserver_core;
-mod key_republisher;
+mod key_republisher_generic;
 mod layers;
 mod periodic_backup;
 mod routes;
